+++ lib/useXhr.ts
@@ -1,66 +1,226 @@
-import React from 'react';
+import * as React from 'react';
+import { v1 as uuidv1 } from 'uuid';
 import {
-  accepts,
-  getFilesFromEvent as defaultGetFilesFromEvent,
+  FileWithPath,
+  fromEvent as defaultGetFilesFromEvent,
+} from 'file-selector';
+import {
+  allFilesAccepted,
+  fileAccepted,
+  fileMatchSize,
   getEstimatedSeconds,
   isIeOrEdge,
+  notEmpty,
+  TOO_MANY_FILES_REJECTION,
 } from './utils';
-import uuidv1 from 'uuid/v1';
+import useEventCallback from './useEventCallback';
 
-const noop = () => {};
+// TODO: add onFileUploaded callback for chunk upload
+// FEATURE?: run in web worker
+// FEATURE: add min image dimensions feature
+// FEATURE: add resume function
 
-function useXhr({
-  getUploadParams = noop,
-  getChunkName = noop,
-  minSizeBytes = 0,
-  maxSizeBytes = Number.MAX_SAFE_INTEGER,
-  maxFiles = Number.MAX_SAFE_INTEGER,
-  accept = '*',
-  autoUpload = true,
-  timeout = 0,
-  validate = noop,
-  noClick = false,
-  multiple = true,
-  disabled = false,
-  chunks = false,
-  chunkSize = 512 * 1024,
-  getFilesFromEvent = defaultGetFilesFromEvent,
-  getDataTransferItemsFromEvent = defaultGetFilesFromEvent,
-} = {}) {
-  const xhrs = React.useRef([]);
-  const mountedRef = React.useRef(true);
-  const inputRef = React.useRef(null);
+type StatusValue =
+  | 'rejected_file_type'
+  | 'rejected_max_files'
+  | 'preparing'
+  | 'error_file_size'
+  | 'error_validation'
+  | 'ready'
+  | 'started'
+  | 'getting_upload_params'
+  | 'error_upload_params'
+  | 'uploading'
+  | 'exception_upload'
+  | 'aborted'
+  | 'restarted'
+  | 'removed'
+  | 'error_upload'
+  | 'headers_received'
+  | 'done';
+
+type MethodValue =
+  | 'delete'
+  | 'get'
+  | 'head'
+  | 'options'
+  | 'patch'
+  | 'post'
+  | 'put'
+  | 'DELETE'
+  | 'GET'
+  | 'HEAD'
+  | 'OPTIONS'
+  | 'PATCH'
+  | 'POST'
+  | 'PUT';
+
+interface RootProps extends React.HTMLAttributes<HTMLElement> {
+  refKey?: string;
+  [key: string]: any;
+}
+
+interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
+  refKey?: string;
+}
+
+type UseXhrProps = {
+  accept?: string;
+  multiple?: boolean;
+  minSizeBytes?: number;
+  maxSizeBytes?: number;
+  maxFiles?: number;
+  autoUpload?: boolean;
+  timeout?: number;
+  chunkSize?: boolean | number;
+  noClick?: boolean;
+  disabled?: boolean;
+  chunks?: boolean;
+  getUploadParams: (file?: FileWithMeta) => UploadParams;
+  getChunkName?: (file: File, chunkIndex: number) => string;
+  getFilesFromEvent?: typeof defaultGetFilesFromEvent;
+  getDataTransferItemsFromEvent?: typeof defaultGetFilesFromEvent;
+  validate?: (
+    file: FileWithPath | DataTransferItem
+  ) => FileError | FileError[] | null;
+  onFileUploaded?(successFile: FileWithMeta, allFiles: FileWithMeta[]): void;
+};
+
+type UploadParams = {
+  url: string;
+  method?: MethodValue;
+  body?: string | FormData | ArrayBuffer | Blob | File | URLSearchParams;
+  fields?: { [name: string]: string | Blob };
+  headers?: { [name: string]: string };
+};
+
+interface Meta {
+  name: string;
+  size: number; // bytes
+  type: string; // MIME type, example: `image/*`
+  lastModifiedDate: string; // ISO string
+  uploadedDate: string; // ISO string
+  progress: number;
+  id: string;
+  status: StatusValue;
+  chunkProgress: (number | 'done')[];
+  estimated: number;
+  previewUrl?: string; // from URL.createObjectURL
+  width?: number;
+  height?: number;
+  duration?: number; // seconds
+  videoWidth?: number;
+  videoHeight?: number;
+  validationError?: Error;
+}
+
+type FileObject = {
+  meta: Meta;
+  cancel: () => void;
+  restart: () => void;
+  remove: () => void;
+};
+
+interface FileWithMeta extends FileObject {
+  file: FileWithPath;
+}
+
+enum ErrorCode {
+  FileInvalidType = 'file-invalid-type',
+  FileTooLarge = 'file-too-large',
+  FileTooSmall = 'file-too-small',
+  TooManyFiles = 'too-many-files',
+}
+
+export interface FileError {
+  message: string;
+  code: ErrorCode | string;
+}
+
+interface FileRejection {
+  file: FileWithPath | DataTransferItem;
+  errors: FileError[];
+}
+
+const useXhr = (
+  {
+    getUploadParams,
+    getChunkName,
+    minSizeBytes = 0,
+    maxSizeBytes = Number.MAX_SAFE_INTEGER,
+    maxFiles = Number.MAX_SAFE_INTEGER,
+    accept = '*',
+    autoUpload = true,
+    validate,
+    timeout,
+    noClick = false,
+    multiple = true,
+    disabled = false,
+    chunks = false,
+    chunkSize = false,
+    getFilesFromEvent = defaultGetFilesFromEvent,
+    getDataTransferItemsFromEvent = defaultGetFilesFromEvent,
+    onFileUploaded = () => {},
+  } = {} as UseXhrProps
+) => {
+  const xhrs = React.useRef<XMLHttpRequest[]>([]);
+  const mountedRef = React.useRef<boolean>(true);
+  const inputRef = React.useRef<HTMLInputElement>(null);
   const activeDragRef = React.useRef(0);
-  const draggedRef = React.useRef([]);
-  const rootRef = React.useRef(null);
+  const draggedRef = React.useRef<(FileWithPath | DataTransferItem)[]>([]);
+  const rootRef = React.useRef<HTMLElement>(null);
   const [isDragActive, setIsDragActive] = React.useState(false);
   const [isFocus, setIsFocus] = React.useState(false);
+  const [rejectedFiles, setRejectedFiles] = React.useState<FileRejection[]>([]);
   const [files, dispatch] = React.useReducer(fileReducer, []);
 
-  const handleUploadError = ({ status }) => {
+  const handleUploadError = ({ status }, id) => {
+    let statusText = 'failure_unknown';
     switch (status) {
       case 404:
         console.error('File not found');
+        statusText = 'failure_file-not-found';
         break;
       case 500:
         console.error('Server error');
+        statusText = 'failure_server-not-found';
         break;
       case 0:
         console.error('Request aborted');
+        statusText = 'failure_aborted';
         break;
       default:
         console.error('Unknown error ' + status);
     }
+    dispatch({
+      type: 'setStatus',
+      payload: { id, status: statusText },
+    });
   };
 
+  function handleResponse({ response, status }, id) {
+    if (status === 201) {
+      dispatch({
+        type: 'setStatus',
+        payload: { id, status: 'done' },
+      });
+      return;
+    }
+    const responseObj = JSON.parse(response);
+    dispatch({
+      type: 'setStatus',
+      payload: { id, status: responseObj.message },
+    });
+  }
+
   const updateFileProgress = useEventCallback(
-    (fileWithMeta, progress, status) => {
+    (fileWithMeta: FileWithMeta, progress: number, status: string) => {
       const fileToUpdate = files.filter(
         file => file.meta.id === fileWithMeta.meta.id
       )[0];
       if (status === 'uploading') {
         dispatch({
-          type: 'UPDATE_PROGRESS',
+          type: 'setProgress',
           payload: {
             id: fileToUpdate.meta.id,
             progress,
@@ -72,77 +232,80 @@ function useXhr({
         });
         return;
       }
-
-      dispatch({
-        type: 'UPDATE_STATUS',
-        payload: { id: fileToUpdate.meta.id, status: 'done' },
-      });
     }
   );
 
-  const allChunksFinished = chunkProgress =>
-    chunkProgress.every(progress => progress === 'done');
-
   const updateFileChunkProgress = useEventCallback(
-    (fileWithMeta, chunkIndex, chunkProgress, status) => {
-      const { id } = fileWithMeta.meta;
+    ({
+      file,
+      chunkIndex,
+      chunkProgress,
+      status,
+    }: {
+      file: FileWithMeta;
+      chunkIndex: number;
+      chunkProgress: number;
+      status: string;
+    }) => {
+      const { id } = file.meta;
       const fileToUpdate = files.filter(file => file.meta.id === id)[0];
       const newChunkProgress = [...fileToUpdate.meta.chunkProgress];
-      const totalProgress =
-        newChunkProgress.reduce(
-          (a, b) =>
-            (typeof a === 'number' ? a : 100) +
-            (typeof b === 'number' ? b : 100),
-          0
-        ) / newChunkProgress.length;
+      const summedProgresses = newChunkProgress.reduce(
+        (a, b) =>
+          (typeof a === 'number' ? a : 100) + (typeof b === 'number' ? b : 100),
+        0
+      ) as number;
+      const totalProgress = summedProgresses / newChunkProgress.length;
       newChunkProgress[chunkIndex] = status === 'done' ? 'done' : chunkProgress;
 
       // Update Progress
       dispatch({
-        type: 'UPDATE_PROGRESS',
+        type: 'setProgress',
         payload: {
           id,
           progress: totalProgress,
           chunkProgress: newChunkProgress,
-          estimated: getEstimatedSeconds(
-            fileWithMeta.meta.uploadedDate,
-            totalProgress
-          ),
+          estimated: getEstimatedSeconds(file.meta.uploadedDate, totalProgress),
         },
       });
-
-      // Set file status to done if all chunks finished uploading
-      if (allChunksFinished(newChunkProgress)) {
-        dispatch({
-          type: 'UPDATE_STATUS',
-          payload: { id, status: 'done' },
-        });
-      }
     }
+  );
+
+  const handleUploadSuccess = React.useCallback(
+    (file: FileWithMeta) => {
+      onFileUploaded(file, files);
+    },
+    [onFileUploaded, files]
   );
 
   // TODO: compare uploadFile and uploadChunk
   // => a lot of duplicate code
   const uploadFile = React.useCallback(
-    (fileWithMeta, progressCallback) => {
-      const fileToUpload = fileWithMeta;
+    ({
+      file,
+      progressCallback,
+    }: {
+      file: FileWithMeta;
+      progressCallback: (progress: any, status: any) => void;
+    }) => {
+      const fileToUpload = file;
       if (!getUploadParams) return;
 
-      let params = null;
+      let params: UploadParams | null = null;
       try {
         params = getUploadParams(fileToUpload);
-      } catch (e) {
+      } catch (e: any) {
         console.error('Error Upload Params', e.stack);
       }
 
       if (params === null || params.url === null) return;
 
       const { url, method = 'POST', body, fields = {}, headers = {} } = params;
-      const { id, size } = fileToUpload.meta;
+      const { id, name, size } = fileToUpload.meta;
 
       if (!url) {
         dispatch({
-          type: 'UPDATE_STATUS',
+          type: 'setStatus',
           payload: { id, status: 'error_upload_params' },
         });
         return;
@@ -154,8 +317,11 @@ function useXhr({
       for (const field of Object.keys(fields)) {
         formData.append(field, fields[field]);
       }
-      xhr.setRequestHeader('X-Content-Length', size);
+      xhr.setRequestHeader('X-Content-Name', name);
+      xhr.setRequestHeader('X-Content-Length', size.toString());
       xhr.setRequestHeader('X-Content-Id', id);
+      xhr.setRequestHeader('X-Chunk-Length', size.toString());
+      xhr.setRequestHeader('X-Chunk-Id', '0');
 
       for (const header of Object.keys(headers)) {
         xhr.setRequestHeader(header, headers[header]);
@@ -164,38 +330,41 @@ function useXhr({
       xhr.onload = () => {
         progressCallback(100, 'done');
         delete xhrs.current[id];
+        handleUploadSuccess(file);
       };
-      xhr.onerror = () => handleUploadError(xhr);
+      xhr.onerror = () => handleUploadError(xhr, id);
+      xhr.onreadystatechange = () =>
+        xhr.readyState === 4 && handleResponse(xhr, id);
       xhr.upload.onprogress = event => {
         if (event.lengthComputable) {
           progressCallback((event.loaded / event.total) * 100, 'uploading');
         }
       };
 
-      formData.append('file', fileToUpload.file);
+      formData.append('chunk', fileToUpload.file);
       if (timeout) xhr.timeout = timeout;
       xhr.send(body || formData);
 
       xhrs.current[id] = xhr;
 
       dispatch({
-        type: 'SET_UPLOADING',
-        payload: id,
+        type: 'setStatus',
+        payload: { id, status: 'uploading' },
       });
     },
-    [getUploadParams, timeout]
+    [getUploadParams, handleUploadSuccess, timeout]
   );
 
   const uploadChunk = React.useCallback(
-    (chunk, chunkIndex, fileWithMeta, progressCallback) => {
+    (chunk, chunkIndex, file, progressCallback) => {
       if (chunk) {
-        const fileToUpload = fileWithMeta;
+        const fileToUpload = file;
         if (!getUploadParams) return;
 
-        let params = null;
+        let params: UploadParams | null = null;
         try {
-          params = getUploadParams();
-        } catch (e) {
+          params = getUploadParams(fileToUpload);
+        } catch (e: any) {
           console.error('Error Upload Params', e.stack);
         }
 
@@ -234,7 +403,9 @@ function useXhr({
           progressCallback(100, chunkIndex, 'done');
           delete xhrs.current[chunkId];
         };
-        xhr.onerror = () => handleUploadError(xhr);
+        xhr.onerror = () => handleUploadError(xhr, id);
+        xhr.onreadystatechange = () =>
+          xhr.readyState === 4 && handleResponse(xhr, id);
         xhr.upload.onprogress = event => {
           if (event.lengthComputable) {
             progressCallback(
@@ -249,11 +420,11 @@ function useXhr({
         if (timeout) xhr.timeout = timeout;
         xhr.send(body || formData);
 
-        xhrs.current[`${fileWithMeta.meta.id}-${chunkId}`] = xhr;
+        xhrs.current[`${file.meta.id}-${chunkId}`] = xhr;
 
         dispatch({
-          type: 'SET_UPLOADING',
-          payload: fileWithMeta.meta.id,
+          type: 'setStatus',
+          payload: { id: file.meta.id, status: 'uploading' },
         });
       }
     },
@@ -261,36 +432,44 @@ function useXhr({
   );
 
   const upload = React.useCallback(
-    fileWithMeta => {
-      if (chunks && typeof chunkSize === 'number') {
-        const BYTES_PER_CHUNK = chunkSize;
-        const SIZE = fileWithMeta.meta.size;
+    (file: FileWithMeta) => {
+      if (chunks) {
+        if (typeof chunkSize === 'number' && chunkSize !== 0) {
+          const BYTES_PER_CHUNK = chunkSize;
+          const SIZE = file.meta.size;
 
-        let start = 0;
-        let end = BYTES_PER_CHUNK;
+          let start = 0;
+          let end = BYTES_PER_CHUNK;
 
-        let chunkIndex = 0;
-        while (start < SIZE) {
-          uploadChunk(
-            fileWithMeta.file.slice(start, end),
-            chunkIndex,
-            fileWithMeta,
-            (percentage, chunkIndex, status) =>
-              updateFileChunkProgress(
-                fileWithMeta,
-                chunkIndex,
-                percentage,
-                status
-              )
+          let chunkIndex = 0;
+          while (start < SIZE) {
+            uploadChunk(
+              file.file.slice(start, end),
+              chunkIndex,
+              file,
+              (percentage, chunkIndex, status) =>
+                updateFileChunkProgress({
+                  file,
+                  chunkIndex,
+                  chunkProgress: percentage,
+                  status,
+                })
+            );
+            chunkIndex += 1;
+            start = end;
+            end = start + BYTES_PER_CHUNK;
+          }
+        } else {
+          console.log(
+            `'chunkSize' has to be a positive number. You provided ${chunkSize}.`
           );
-          chunkIndex += 1;
-          start = end;
-          end = start + BYTES_PER_CHUNK;
         }
       } else {
-        uploadFile(fileWithMeta, (progress, status) =>
-          updateFileProgress(fileWithMeta, progress, status)
-        );
+        uploadFile({
+          file,
+          progressCallback: (progress, status) =>
+            updateFileProgress(file, progress, status),
+        });
       }
     },
     [
@@ -303,7 +482,8 @@ function useXhr({
     ]
   );
 
-  const getFileById = id => files.filter(file => file.meta.id === id)[0];
+  const getFileById = (id: string) =>
+    files.filter(file => file.meta.id === id)[0];
 
   const abortUpload = fileToAbort => {
     const { id, chunkProgress } = fileToAbort.meta;
@@ -324,7 +504,7 @@ function useXhr({
     if (fileToCancel && fileToCancel.meta.status === 'uploading') {
       abortUpload(fileToCancel);
       dispatch({
-        type: 'CANCEL_UPLOAD',
+        type: 'cancelUpload',
         payload: id,
       });
     }
@@ -356,11 +536,16 @@ function useXhr({
         abortUpload(fileToRemove);
       }
       dispatch({
-        type: 'REMOVE_FILE_BY_ID',
+        type: 'removeById',
         payload: id,
       });
     }
   });
+
+  const uploadAll = React.useCallback(() => {
+    if (!files.length) return;
+    files.forEach(restart);
+  }, [files, restart]);
 
   const handleFile = React.useCallback(
     async (file, id) => {
@@ -368,7 +553,7 @@ function useXhr({
       const uploadedDate = new Date().toISOString();
       const lastModifiedDate =
         lastModified && new Date(lastModified).toISOString();
-      let chunkProgress = null;
+      let chunkProgress: number[];
 
       if (typeof chunkSize === 'number') {
         chunkProgress = [];
@@ -385,42 +570,17 @@ function useXhr({
           type,
           lastModifiedDate,
           uploadedDate,
+          percent: 0,
           progress: 0,
-          chunkProgress,
-          id,
           estimated: 0,
+          chunkProgress: [],
+          id,
+          status: 'preparing',
         },
-      };
-
-      // Firefox versions prior to 53 return a bogus mime type for file drag events,
-      // so files with that mime type are always accepted
-      if (file.type !== 'application/x-moz-file' && !accepts(file, accept)) {
-        fileWithMeta.meta.status = 'rejected_file_type';
-        dispatch({ type: 'ADD_FILE', payload: fileWithMeta });
-        return;
-      }
-
-      if (files.length >= maxFiles) {
-        fileWithMeta.meta.status = 'rejected_max_files';
-        dispatch({ type: 'ADD_FILE', payload: fileWithMeta });
-        return;
-      }
-
-      if (size < minSizeBytes || size > maxSizeBytes) {
-        fileWithMeta.meta.status = 'error_file_size';
-        dispatch({ type: 'ADD_FILE', payload: fileWithMeta });
-        return;
-      }
-
-      if (validate) {
-        const error = validate(fileWithMeta);
-        if (error) {
-          fileWithMeta.meta.status = 'error_validation';
-          fileWithMeta.meta.validationError = error;
-          dispatch({ type: 'ADD_FILE', payload: fileWithMeta });
-          return;
-        }
-      }
+        cancel: () => {},
+        remove: () => {},
+        restart: () => {},
+      } as FileWithMeta;
 
       fileWithMeta.meta.status = autoUpload ? 'getting_upload_params' : 'ready';
 
@@ -430,33 +590,69 @@ function useXhr({
 
       const fileWithMetaAndPreview = await generatePreview(fileWithMeta);
 
-      dispatch({ type: 'ADD_FILE', payload: fileWithMetaAndPreview });
+      dispatch({ type: 'add', payload: fileWithMetaAndPreview });
       if (autoUpload) {
         upload(fileWithMetaAndPreview);
       }
     },
-    [
-      chunkSize,
-      accept,
-      files.length,
-      maxFiles,
-      minSizeBytes,
-      maxSizeBytes,
-      validate,
-      autoUpload,
-      cancel,
-      remove,
-      restart,
-      upload,
-    ]
+    [chunkSize, autoUpload, cancel, remove, restart, upload]
   );
 
-  // Expects an array of File objects
   const handleFiles = React.useCallback(
-    addedFiles => {
-      addedFiles.forEach(file => handleFile(file, uuidv1()));
+    (addedFiles: (FileWithPath | DataTransferItem)[]) => {
+      setRejectedFiles([]);
+      const acceptedFiles: (FileWithPath | DataTransferItem)[] = [];
+      const fileRejections: FileRejection[] = [];
+
+      addedFiles.forEach(file => {
+        const [accepted, acceptError] = fileAccepted(file, accept);
+        const [sizeMatch, sizeError] = fileMatchSize(
+          file,
+          minSizeBytes,
+          maxSizeBytes
+        );
+        const customErrors = validate ? validate(file) : null;
+
+        if (accepted && sizeMatch && !customErrors) {
+          acceptedFiles.push(file);
+        } else {
+          let errors = [acceptError, sizeError];
+
+          if (customErrors) {
+            errors = errors.concat(customErrors);
+          }
+
+          fileRejections.push({ file, errors: errors.filter(notEmpty) });
+        }
+
+        if (
+          (!multiple && acceptedFiles.length > 1) ||
+          (multiple && maxFiles >= 1 && acceptedFiles.length > maxFiles) ||
+          // if autoUpload is disabled, we want to check the length of all files in the queue
+          (!autoUpload && files.length + acceptedFiles.length > maxFiles)
+        ) {
+          // Reject everything and empty accepted files
+          acceptedFiles.forEach(file => {
+            fileRejections.push({ file, errors: [TOO_MANY_FILES_REJECTION] });
+          });
+          acceptedFiles.splice(0);
+        }
+      });
+
+      acceptedFiles.forEach(file => handleFile(file, uuidv1()));
+      setRejectedFiles(fileRejections);
     },
-    [handleFile]
+    [
+      autoUpload,
+      files,
+      accept,
+      handleFile,
+      maxFiles,
+      maxSizeBytes,
+      minSizeBytes,
+      multiple,
+      validate,
+    ]
   );
 
   const openFileDialog = () => {
@@ -525,7 +721,7 @@ function useXhr({
 
     // TODO: test in IE if we need this
     // https://github.com/react-dropzone/react-dropzone/blob/master/src/index.js#L465
-    // In IE11/Edge the file-browser dialog is blocking, therefore, use set = 0()
+    // In IE11/Edge the file-browser dialog is blocking, therefore, use setTimeout()
     if (isIeOrEdge()) {
       setTimeout(openFileDialog, 0);
     } else {
@@ -534,14 +730,13 @@ function useXhr({
   }, [noClick]);
 
   const onKeyDown = React.useCallback(
-    event => {
+    (event: React.BaseSyntheticEvent<KeyboardEvent, HTMLDivElement>) => {
       // Ignore keyboard events bubbling up the DOM tree
-      if (!rootRef.current || !rootRef.current.isEqualNode(event.target)) {
+      if (!rootRef.current || !rootRef.current.isEqualNode(event.target))
         return;
-      }
 
       // Open file dialog with 'space' and 'enter' keys
-      if (event.keyCode === 32 || event.keyCode === 13) {
+      if (event.nativeEvent.key === ' ' || event.nativeEvent.key === 'Enter') {
         event.preventDefault();
         openFileDialog();
       }
@@ -554,12 +749,14 @@ function useXhr({
       meta: { type },
       file,
     } = fileWithMeta;
-    const preview = {};
+    const preview: any = {};
     const isImage = type.startsWith('image/');
     const isAudio = type.startsWith('audio/');
     const isVideo = type.startsWith('video/');
+    const isApplication = type.startsWith('application/');
+    const isText = type.startsWith('text/');
 
-    if (!isImage && !isAudio && !isVideo) return;
+    if (!isImage && !isAudio && !isVideo && !isApplication && !isText) return;
 
     const objectUrl = URL.createObjectURL(file);
     const fileCallbackToPromise = fileObj => {
@@ -609,9 +806,18 @@ function useXhr({
     };
   };
 
-  const isDragReject = draggedRef.current.some(
-    file => file.type !== 'application/x-moz-file' && !accepts(file, accept)
-  );
+  const fileCount = draggedRef.current.length;
+  const isDragAccept =
+    fileCount > 0 &&
+    allFilesAccepted({
+      files: draggedRef.current,
+      accept,
+      minSize: minSizeBytes,
+      maxSize: maxSizeBytes,
+      multiple,
+      maxFiles,
+    });
+  const isDragReject = fileCount > 0 && !isDragAccept;
 
   const extra = {
     accept,
@@ -622,40 +828,42 @@ function useXhr({
   };
 
   const getRootProps = React.useMemo(
-    () => ({ refKey = 'ref', ...rest } = {}) => ({
-      onDragEnter,
-      onDragOver,
-      onDragLeave,
-      onFocus,
-      onBlur,
-      onKeyDown,
-      onClick,
-      onDrop: !disabled ? onDrop : onDropDisabled,
-      ...(!disabled ? { tabIndex: 0 } : {}),
-      [refKey]: rootRef,
-      ...rest,
-    }),
+    () =>
+      ({ refKey = 'ref', ...rest }: RootProps = {}) => ({
+        onDragEnter,
+        onDragOver,
+        onDragLeave,
+        onFocus,
+        onBlur,
+        onKeyDown,
+        onClick,
+        onDrop: !disabled ? onDrop : onDropDisabled,
+        ...(!disabled ? { tabIndex: 0 } : {}),
+        [refKey]: rootRef,
+        ...rest,
+      }),
     [onDragEnter, onDragOver, onDragLeave, onKeyDown, onClick, disabled, onDrop]
   );
 
   const getInputProps = React.useMemo(
-    () => ({ refKey = 'ref', onChange, ...rest } = {}) => ({
-      accept,
-      multiple,
-      type: 'file',
-      style: { display: 'none' },
-      onChange: async event => {
-        event.persist();
-        const chosenFiles = await getFilesFromEvent(event);
-        handleFiles(chosenFiles);
-        event.target.value = null;
-      },
-      autoComplete: 'off',
-      tabIndex: -1,
-      disabled: disabled !== undefined ? disabled : noClick,
-      [refKey]: inputRef,
-      ...rest,
-    }),
+    () =>
+      ({ refKey = 'ref', onChange, ...rest }: InputProps = {}) => ({
+        accept,
+        multiple,
+        type: 'file',
+        style: { display: 'none' },
+        onChange: async event => {
+          event.persist();
+          const chosenFiles = await getFilesFromEvent(event);
+          handleFiles(chosenFiles);
+          event.target.value = null;
+        },
+        autoComplete: 'off',
+        tabIndex: -1,
+        disabled: disabled !== undefined ? disabled : noClick,
+        [refKey]: inputRef,
+        ...rest,
+      }),
     [accept, multiple, disabled, noClick, getFilesFromEvent, handleFiles]
   );
 
@@ -671,28 +879,36 @@ function useXhr({
 
   return {
     files,
+    rejectedFiles,
     extra,
     getRootProps,
     getInputProps,
     isDragActive,
+    isDragAccept,
     isDragReject,
     isFocus,
     inputRef,
+    uploadAll,
   };
+};
+
+interface Action {
+  type: string;
+  payload: any;
 }
 
-function fileReducer(files, action) {
+function fileReducer(files: FileWithMeta[], action: Action): FileWithMeta[] {
   switch (action.type) {
-    case 'ADD_FILE': {
+    case 'add': {
       return [...files, action.payload];
     }
-    case 'REMOVE_FILE_BY_ID': {
+    case 'removeById': {
       return files.filter(file => file.meta.id !== action.payload);
     }
-    case 'RESET': {
+    case 'reset': {
       return [];
     }
-    case 'CANCEL_UPLOAD': {
+    case 'cancelUpload': {
       return files.map(file => {
         if (file.meta.id !== action.payload) {
           return file;
@@ -707,7 +923,7 @@ function fileReducer(files, action) {
         }
       });
     }
-    case 'UPDATE_STATUS': {
+    case 'setStatus': {
       return files.map(file => {
         if (file.meta.id !== action.payload.id) {
           return file;
@@ -722,7 +938,7 @@ function fileReducer(files, action) {
         }
       });
     }
-    case 'UPDATE_PROGRESS': {
+    case 'setProgress': {
       return files.map(file => {
         if (file.meta.id !== action.payload.id) {
           return file;
@@ -739,36 +955,10 @@ function fileReducer(files, action) {
         }
       });
     }
-    case 'SET_UPLOADING': {
-      return files.map(file => {
-        if (file.meta.id !== action.payload) {
-          return file;
-        } else {
-          return {
-            ...file,
-            meta: {
-              ...file.meta,
-              status: 'uploading',
-            },
-          };
-        }
-      });
-    }
     default:
       return files;
   }
 }
 
-// NOTE: Gives us the current state without invalidating too often and
-// let us do side effects
-// https://github.com/facebook/react/issues/14099#issuecomment-440013892
-// but have in mind: https://github.com/facebook/react/issues/14092#issuecomment-435907249
-function useEventCallback(fn) {
-  let ref = React.useRef();
-  React.useLayoutEffect(() => {
-    ref.current = fn;
-  });
-  return React.useCallback((...args) => (0, ref.current)(...args), []);
-}
-
-export default useXhr;
+export { ErrorCode } from './utils';
+export { useXhr };