+++ lib/utils.ts
@@ -1,0 +1,147 @@
+import accepts from 'attr-accept';
+import { FileError } from './useXhr';
+
+export const FILE_INVALID_TYPE = 'file-invalid-type';
+export const FILE_TOO_LARGE = 'file-too-large';
+export const FILE_TOO_SMALL = 'file-too-small';
+export const TOO_MANY_FILES = 'too-many-files';
+
+export const ErrorCode = {
+  FileInvalidType: FILE_INVALID_TYPE,
+  FileTooLarge: FILE_TOO_LARGE,
+  FileTooSmall: FILE_TOO_SMALL,
+  TooManyFiles: TOO_MANY_FILES,
+};
+
+export function getEstimatedSeconds(starttime: string, progress: number) {
+  const timePassed = Date.now() - +new Date(starttime);
+  const estimated = (100 - progress) * (timePassed / progress);
+  const estimatedSeconds = Math.floor((estimated / 1000) % 60);
+  return estimatedSeconds || 0;
+}
+
+export const getInvalidTypeRejectionErr = (accept: string): FileError => {
+  accept = Array.isArray(accept) && accept.length === 1 ? accept[0] : accept;
+  const messageSuffix = Array.isArray(accept)
+    ? `one of ${accept.join(', ')}`
+    : accept;
+  return {
+    code: FILE_INVALID_TYPE,
+    message: `File type must be ${messageSuffix}`,
+  };
+};
+
+export const getTooLargeRejectionErr = maxSize => {
+  return {
+    code: FILE_TOO_LARGE,
+    message: `File is larger than ${maxSize} bytes`,
+  };
+};
+
+export const getTooSmallRejectionErr = minSize => {
+  return {
+    code: FILE_TOO_SMALL,
+    message: `File is smaller than ${minSize} bytes`,
+  };
+};
+
+export const TOO_MANY_FILES_REJECTION = {
+  code: TOO_MANY_FILES,
+  message: 'Too many files',
+};
+
+// Firefox versions prior to 53 return a bogus MIME type for every file drag, so dragovers with
+// that MIME type will always be accepted
+export function fileAccepted(
+  file,
+  accept: string
+): [boolean, null | FileError] {
+  const isAcceptable =
+    file.type === 'application/x-moz-file' ||
+    file.type === '' || // accept empty string, it could be a directory
+    accepts(file, accept);
+  return [
+    isAcceptable,
+    isAcceptable ? null : getInvalidTypeRejectionErr(accept),
+  ];
+}
+
+export function fileMatchSize(
+  file,
+  minSize: number,
+  maxSize: number
+): [boolean, null | FileError] {
+  if (isDefined(file.size)) {
+    if (isDefined(minSize) && isDefined(maxSize)) {
+      if (file.size > maxSize) return [false, getTooLargeRejectionErr(maxSize)];
+      if (file.size < minSize) return [false, getTooSmallRejectionErr(minSize)];
+    } else if (isDefined(minSize) && file.size < minSize)
+      return [false, getTooSmallRejectionErr(minSize)];
+    else if (isDefined(maxSize) && file.size > maxSize)
+      return [false, getTooLargeRejectionErr(maxSize)];
+  }
+  return [true, null];
+}
+
+function isDefined(value) {
+  return value !== undefined && value !== null;
+}
+
+export function allFilesAccepted({
+  files,
+  accept,
+  minSize,
+  maxSize,
+  multiple,
+  maxFiles,
+}) {
+  if (
+    (!multiple && files.length > 1) ||
+    (multiple && maxFiles >= 1 && files.length > maxFiles)
+  ) {
+    return false;
+  }
+
+  return files.every(file => {
+    const [accepted] = fileAccepted(file, accept);
+    const [sizeMatch] = fileMatchSize(file, minSize, maxSize);
+    return accepted && sizeMatch;
+  });
+}
+
+// TODO: not used right now
+export function isEvtWithFiles(event) {
+  if (!event.dataTransfer) {
+    return !!event.target && !!event.target.files;
+  }
+  // https://developer.mozilla.org/en-US/docs/Web/API/DataTransfer/types
+  // https://developer.mozilla.org/en-US/docs/Web/API/HTML_Drag_and_Drop_API/Recommended_drag_types#file
+  return Array.prototype.some.call(
+    event.dataTransfer.types,
+    type => type === 'Files' || type === 'application/x-moz-file'
+  );
+}
+
+function isIe(userAgent: string) {
+  return (
+    userAgent.indexOf('MSIE') !== -1 || userAgent.indexOf('Trident/') !== -1
+  );
+}
+
+function isEdge(userAgent: string) {
+  return userAgent.indexOf('Edge/') !== -1;
+}
+
+export function isIeOrEdge(userAgent: string = window.navigator.userAgent) {
+  return isIe(userAgent) || isEdge(userAgent);
+}
+
+export function notEmpty<TValue>(
+  value: TValue | null | undefined
+): value is TValue {
+  if (value === null || value === undefined) return false;
+  // eslint-disable-next-line
+  const testDummy: TValue = value;
+  return true;
+}
+