+++ setup.tests.js
@@ -1,0 +1,3 @@
+// this adds custom jest matchers from jest-dom.
+// More here: https://github.com/gnapse/jest-dom
+import '@testing-library/jest-dom/extend-expect';