+++ CHANGELOG.md
@@ -2,4 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### 0.0.2 (2019-08-25)
+### [0.1.3](https://gitlab.3pc.de/frontend/packages/use-xhr/compare/v0.1.2...v0.1.3) (2021-12-01)
+
+
+### Bug Fixes
+
+* default exporting ([27852d9](https://gitlab.3pc.de/frontend/packages/use-xhr/commit/27852d9a7ab011105951a29b634d8989013a8417))
+
+### [0.1.2](https://gitlab.3pc.de/frontend/packages/use-xhr/compare/v0.1.1...v0.1.2) (2021-12-01)
+
+### 0.1.1 (2021-12-01)
+
+
+### Bug Fixes
+
+* testing and setup ([d1fca0c](https://gitlab.3pc.de/frontend/packages/use-xhr/commit/d1fca0ce960d04997aaa9dc942b0ea9bb04a1919))